requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
gspread>=5.11.0
google-auth>=2.23.0
openpyxl>=3.1.0
plotly>=5.17.0
